Add rootfs tar entry. Don't use aci.ArchiveWriter
The ACI format expects a rootfs entry in the tar file, we now add one
when generating the ACIs.

Also, aci.ArchiveWriter writes the manifest with current date, which
leads to non-reproducable ACIs. I don't think modifying the aci package
is the right thing to do so we write directly to the tar file.