Remove docker dependencies
We depended on docker packages to use docker's json manifest structures
and to parse the registry URLs but they're in packages `parsers`,
`registry` and `runconfig` which are quite general and have a lot of
dependencies. This caused Godep to download a big chunk of docker's code
which is not desirable when integrating docker2aci with rocket, for
example.

This commit extracts the needed docker functions and structures so we
don't import docker packages anymore.