Support local images
Adds support for converting Docker images generated by "docker save".
Since this file can include several images and tags, we convert whatever
image is there if there's only one or we politely ask the user to select
one image if there're several.

This also changes docker2aci's interface. We now require the prefix
"docker://" to get images from a Docker registry and assume that if it's
not present the user means a local file.