Handle whiteout files
We were ignoring Docker's whiteout files so it's time to handle them.

To do that, we keep a white list of files (PathWhitelist). As we go
through the layers, we remove any files maked as whiteouts from the list
so so that, at any time, each layer ACI has the list of its desired
files and the ones of its parent layers.

We change order we process the images to start from the base image so
that we can generate the correct white lists. Then we reverse the image
list because acirenderer expects images starting from the upper layer.