Fix retrieving images from Google Container Registry
Google Container Registry doesn't implement the Docker registry API
correctly: when trying to access
"REGISTRY/repositories/IMAGE/NAME/tags/TAG" it fails with a 405.

Instead of getting the tag from that URL, get the JSON with all the tags
and look for the one we want. This is what Docker does.