docker2aci: convert Docker Volumes to mountPoints
A Docker image can specify Volumes which are akin to mountPoints in the
appc spec. This commit converts Volumes to mountPoints allowing the user
to mount volumes with rkt's --volume option.

Since Docker doesn't name Volumes but the appc spec requires a name, we
generate a name by appending the path to "volume-". That is, if a Volume
has "/var/tmp" as path, the resulting mountPoint name will be
"volume-/var/tmp". A mountPoint name must be a valid ACName so any
invalid characters will be replaced by a "-".