lib: properly write squashed ACI to disk
We were writing the file directly to its final destination instead of
writing it to a temporary file and then renaming it. This caused file
corruption when running docker2aci several times in parallel.

We now write the squashed ACI to a temporary file with a random name and
rename it atomically after the write is finished and the ACI validated.