lib/common: make a copy of pwl instead of reusing the same var
In f78ecc1d6e7c3fd913b0d374f759a5b6d3a628f2, we started sorting every
golang map before serializing the Image Manifests to make the generated
ACIs deterministic (golang randomizes maps).

This causes problems with squashed ACIs: they're missing some files (see
https://github.com/coreos/rkt/issues/2244 for some examples).

The reason for this is that we were modifying the PathWhitelist inside
writeACI() and then reusing the same copy in the next iteration of the
loop in convertReal(). Making a copy of the modified PathWhitelist when
we generate the layer's manifest fixes the problem.