lib/repository2: add support for docker v2.2 and OCI
This commit adds support for the docker v2.2 image format and OCI image
format. When the remote registry supports the docker v2 HTTP API,
docker2aci will add the media types for v2.1, v2.2 and OCI to the Accept
header on the manifest request, and handle the returned manifest version
accordingly.