repository2: get the correct layer index
In da56c93f we introduced an optimization to avoid going through all the
layers every time we need the index of a specific one.

However, this resulted in incorrect Image Manifests being generated.

In f8e82fab we attempted to fix this. We assumed that the mistake was
that the index we used was supposed to be reversed (from the end instead
of from the beginning of the layer list). This assumption was wrong. and
was probably motivated by the name of the map (`reverseLayers`).

The problem was that there can be several layers with the same BlobSum
(layerID) but different manifest and when we populated the map, we were
storing the last one for a given BlobSum. This is not what we were doing
previously, instead we were returning the first match.

Fix this by storing the first appearance of a BlobSum. Also, change the
name of `reverseLayers` to `layersIndex`.