common: move types.ParseDockerURL to common.ParseDockerURL
The type common.ParsedDockerURL is aliased to types.ParsedDockerURL
which is unnecessary. This removes the alias, and moves the type from
the internal `types` package to the `common` package.

Additionally common.ParseDockerURL is only delegating to
types.ParseDockerURL which is unnecessary. This moves the implementation
from the internal `types` package to the `common` package.