file: treat file vs remote more consistently
This has multiple side effects, but the most obvious way to see this is
more correct is to compare the following:

$ docker2aci docker://busybox
...
Generated ACI(s):
library-busybox-latest.aci

$ docker pull busybox && docker save busybox > busybox.tar
$ docker2aci ./busybox.tar
...
Generated ACI(s):
busybox-latest.aci

Note the difference in output filename. This change causes them to both
be the former.

This becomes a bigger problem, however, in trying to use the `-image`
flag feature.

Prior to this change, the image flag with the above tarball would fail
with:

$ docker2aci --image="busybox" ./busybox.tar
Error: conversion error: app "library/busybox" not found

After this change, it operates as expected on the above example.