lib: stop downloading images if the caller already has this manifest
This commit allows a user of the library to provide a list of manifest
hashes, presumably generated from annotations left on images by prior
invocations of docker2aci, for images it already has. If docker2aci
fetches a manifest and the hash for it matches one of the elements in
the user provided list, docker2aci will stop working, and return with no
error and no produced ACIs.