Avoid panicing on scratch images
Docker allows creating empty images (e.g. with FROM scratch followed by
a no-op like MAINTAINER or LABEL).

Without this patch, docker2aci will panic when it encounters an image of
this sort.

These images are more-or-less useless. They're also difficult to work
with. As an example, `docker save` bails on them with an `unimplemented`
error.

In line with that behavior, I think bailing in docker2aci makes sense
too rather than dealing with more robust error handling or creating an
empty ACI.